% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\docType{methods}
\name{clust.accessors}
\alias{clust.accessors}
\alias{clustCenters}
\alias{clustCenters,clust-method}
\alias{clustCluster}
\alias{clustCluster,clust-method}
\alias{clustData}
\alias{clustData,clust-method}
\alias{clustMembership}
\alias{clustMembership,clust-method}
\title{Accessors to extract slots of a clust class.}
\usage{
clustData(object)

\S4method{clustData}{clust}(object)

clustCenters(object)

\S4method{clustCenters}{clust}(object)

clustCluster(object)

\S4method{clustCluster}{clust}(object)

clustMembership(object)

\S4method{clustMembership}{clust}(object)
}
\arguments{
\item{object}{\code{clust} object object}
}
\value{
\code{clustData} returns data matrix. \code{clustCenters} returns a matrix of
centers. \code{clustCluster} returns an integer vector. \code{clustMembership}
returns a matrix of membership, see \code{\link{clust}} for details.
}
\description{
Accessors are provided to extract \code{data}, \code{centers}, \code{cluster},
\code{membership}, \code{membership} slots of a clust class.
}
\author{
Mengjun Wu
}
\seealso{
\code{\link{clust}}
}

