% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countsTable}
\alias{countsTable}
\title{An example read Counts table}
\format{A data frame containing experiment design information
for 12 samples/libraries.}
\usage{
data(countsTable)
}
\value{
A data frame
}
\description{
A dataset of exemplary read counts
}
\examples{
data(countsTable)
}
\keyword{datasets}

