% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDBresult.R
\name{DBresult.filter}
\alias{DBresult.filter}
\title{Extract significant differential events
This function extracts genomic regions that have significant differential events by log2-fold changes,p-values or adjusted p-values.
Features passing certain thresholds will be kept while other features will be dropped.}
\usage{
DBresult.filter(x, pvalue = "paj", pvalue.threshold = 0.05, abs.fold = 2,
  direction = "both")
}
\arguments{
\item{x}{list of data frames of differential binding results (such as returned value of \code{\link{DBresult}}),
 required columns of the data frames are
\code{logFC}, \code{PValue}, \code{paj}.}

\item{pvalue}{character string specify the type of p-values ('\code{PValue}' or adjusted p-value 'paj')}

\item{pvalue.threshold}{a numeric value giving threshold of selected p-value, only features with higher
(ajusted) p-values than the threshold are kept.}

\item{abs.fold}{a numeric value, the least absolute log2-fold changes}

\item{direction}{character string specify the direction of fold changes ('\code{up}' (positive fold changes),
'\code{down}' (negative fold changes), \code{both}'(both positive and negative fold changes)), features changes
more than \code{abs.fold} in the defined direction are kept.}
}
\value{
A list of data frames
}
\description{
Extract significant differential events
This function extracts genomic regions that have significant differential events by log2-fold changes,p-values or adjusted p-values.
Features passing certain thresholds will be kept while other features will be dropped.
}
\examples{
data(tca_ATAC)
tca_ATAC <- DBanalysis(tca_ATAC)
res <- DBresult(tca_ATAC, group1 = '0h', group2 = c('24h', '72h'))
top_res <- DBresult.filter(res)

}
\author{
Mengjun Wu, Lei Gu
}
\seealso{
\code{\link{DBresult}}
}

