% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBresults.R
\name{DBresult}
\alias{DBresult}
\alias{DBresult.cluster}
\title{This function performs differetial analysis by fitting read
Perform likelihood ratio tests and extract the differential
analysis results}
\usage{
DBresult(object, group1 = NULL, group2 = NULL, contrasts = NULL,
  p.adjust = "fdr", top.sig = FALSE, pvalue = "paj",
  pvalue.threshold = 0.05, abs.fold = 2, direction = "both",
  result.type = "GRangesList")

DBresult.cluster(object, group1 = NULL, group2 = NULL, contrasts = NULL,
  p.adjust = "fdr", top.sig = FALSE, pvalue = "paj",
  pvalue.threshold = 0.05, abs.fold = 2, direction = "both", cluster,
  cmthreshold = NULL, result.type = "GRangesList")
}
\arguments{
\item{object}{a \code{TCA} object, for \code{DBresult},
\code{DBanalysis} should already be called on the object;
for \code{DBresult.cluster}, both \code{DBanalysis} and
\code{timeclust} should be already called.}

\item{group1}{character string giving the level to be compared,
that is the denominator in the fold changes.}

\item{group2}{a character vetor giving other levels to compared
with \code{group1}. that are numerator in the fold changes.}

\item{contrasts}{a character vector, each charcter string in
the vector gives a contrast of two groups with the format
group2vsgroup1', group1 is the denominator level in the fold
changes and group2 is the numerator
level in the fold changes.}

\item{p.adjust}{character string specifying a correction method
for p-values. Options are 'holm', hochberg', 'hommel',
'bonferroni', BH', 'BY', 'fdr', 'none'.}

\item{top.sig}{logical if TRUE, only genomic regions with
significant differential events will are returned. Significant
differential events are defined by log2-fold changes,p-values or
adjusted p-values.}

\item{pvalue}{character string specify the type of p-values
used to define significant differential events('\code{PValue}'
or adjusted p-value 'paj')}

\item{pvalue.threshold}{a numeric value giving threshold of
selected p-value, Significant differential events have lower
(ajusted) p-values than the threshold.}

\item{abs.fold}{a numeric value, the least absolute log2-fold
changes}

\item{direction}{character string specify the direction of fold
changes ('\code{up}' (positive fold changes), \code{down}'
(negative fold changes), \code{both}'(both positive and
negative fold changes)). Significant events have log2-fold
changes exceeding \code{abs.fold} in defined directions.}

\item{result.type}{character string giving the data type of return
value. Options are "GRangesList" and "list".}

\item{cluster}{an integer, the result tables of genomic features
belong to the \code{cluster} are extracted.}

\item{cmthreshold}{a numeric value, this argument is applicable
only if \code{cmeans}' clustering method is selected when calling
\code{\link{timeclust}} function. if not NULL, the result table of
genomic features that belong to the defined \code{cluster} and
the membership values to this cluster exceed \code{cmthreshold}
are extracted.}
}
\value{
A list or a GRangesList.
If \code{result.type} is "GRangesList", a GRangesList is returned
containing the differential analysis results for all provided contrasts.
Each GRanges object of the list is one contrast, the analysis results
are contained in 4 metadata columns:

\code{logFC} log2-fold changes of differential event between
two tested.

\code{PValue} p-values.

\code{paj} adjusted p-values

\code{id} genomic feature name

If \code{result.type} is "list", a List of data frames is returned.
Each data frame is one contrast and contains the following columns:

\code{logFC} log2-fold changes of differential event between
two tested.

\code{PValue} p-values.

\code{paj} adjusted p-values

\code{chr}  name of the chromosomes

\code{start} starting position of the feature in the
chromosome

\code{end} ending postition of the feature in the chromosome

\code{id} genomic feature name
}
\description{
This function performs likelihood ratio tests for given
coefficinets contrasts after fitting read counts to GLM by
\code{\link{DBanalysis}}. \code{DBresult} extracts the
diffential analysis results of given contrasts for all
genomic features or genomic features with significant
differential events. \code{DBresult.cluster} returns similar
results while the results only contain genomic features belong
to a given cluster.
}
\details{
This function uses \code{\link{glmLRT}} from edgeR which
perform likelihood ratio tests for testing significance of changes.
For more deatils,
see \code{\link{glmLRT}}
}
\note{
If not NULL \code{group1}, \code{group2} and \code{contrasts},
result tables are extracted from comparisons in \code{constrasts}.
}
\examples{
data(tca_ATAC)
tca_ATAC <- DBanalysis(tca_ATAC)
### extract differntial analysis of 24h, 72h to 0h
# set the contrasts using the 'group1' and 'group2' paramters
res1 <- DBresult(tca_ATAC, group1 = '0h', group2 = c('24h', '72h'))
# one can get the same result by setting the contrasts using hte 'contrasts' parameter
res2 <- DBresult(tca_ATAC, contrasts = c('24hvs0h', '72hvs0h'))
# extract significant diffential events
res.sig <- DBresult(tca_ATAC, contrasts = c('24hvs0h', '72hvs0h'),
                   top.sig = TRUE)

# extract differntial analysis of 24h, 72h to 0h of a given cluster
tca_ATAC <- timecourseTable(tca_ATAC, filter = TRUE)
tca_ATAC <- timeclust(tca_ATAC, algo = 'cm', k = 6)
res_cluster1 <- DBresult.cluster(tca_ATAC, group1 = '0h',
                                 group2 = c('24h', '72h'),
                                 cluster = 1)



}
\author{
Mengjun Wu, Lei Gu
}
\seealso{
\code{\link{glmLRT}}
}

