% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment}
\alias{experiment}
\title{An example experiment design without BAM file infomration}
\format{
A data frame containing experiment design information
for 12 samples/libraries.
}
\usage{
data(experiment)
}
\value{
A data frame
}
\description{
A dataset of exemplary experiment design without BAM file
infomration
}
\examples{
data(experiment)
}
\keyword{datasets}
