% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_BAMfile}
\alias{experiment_BAMfile}
\title{An example experiment design with BAM file infomration}
\format{
A data frame containing experiment design information
for 12 samples/libraries.
}
\usage{
data(experiment_BAMfile)
}
\value{
A data frame
}
\description{
A dataset of exemplary experiment design with BAM file
infomration
}
\examples{
data(experiment_BAMfile)
}
\keyword{datasets}
