% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genomicIntervals}
\alias{genomicIntervals}
\title{An example reference genomic regions}
\format{A data frame containing 2751 genomic regions.}
\usage{
data(genomicIntervals)
}
\value{
A data frame
}
\description{
A dataset of exemplary genomic regions
}
\examples{
data(genomicIntervals)
}
\keyword{datasets}

