% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tca_ATAC}
\alias{tca_ATAC}
\title{An example TCA object}
\format{
A TCA object of exemplary ATAC-seq time course data
}
\usage{
data(tca_ATAC)
}
\value{
A TCA object
}
\description{
A TCA object storing exemplary ATAC-seq time course data,
including the experiment design, read counts, reference
genomic regions.
}
\examples{
data(tca_ATAC)
}
\keyword{datasets}
