% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourseTable.R
\name{timecourseTable}
\alias{timecourseTable}
\title{constructs time course table for clustering analysis}
\usage{
timecourseTable(object, value = "expression", lib.norm = TRUE,
  norm.method = "rpkm", subset = NULL, filter = FALSE, pvalue = "fdr",
  pvalue.threshold = 0.05, abs.fold = 2, direction = "both", ...)
}
\arguments{
\item{object}{a \code{TCA} object returned by \code{DBanalysis}.}

\item{value}{character string, either '\code{expression}' or '\code{FC}'. '\code{expression}' is the
mean normalized read counts of replicates, '\code{FC}' is the log2-fold changes compared to the first
time point.}

\item{lib.norm}{logical indicating whether or not use effective library size
(see 'Details' in \code{\link{counts}}).}

\item{norm.method}{character string specifying the normalization method if \code{value} is '\code{expression}'}

\item{subset}{optinal character vector giving a subset of genomic features, if not NULL, time course table is generated
for only this subset of genomic features.}

\item{filter}{logical, whether to drop the genomic features shows no significant changes (defined by \code{pvalue},
\code{pvalue.threshold}, \code{abs.fold} and \code{direction}) between any two time points.}

\item{pvalue}{character string specify the type of p-values ('\code{PValue}' or adjusted p-value 'paj')}

\item{pvalue.threshold}{a numeric value giving threshold of selected p-value, only features with higher
(ajusted) p-values than the threshold are kept.}

\item{abs.fold}{a numeric value, the least absolute log2-fold changes}

\item{direction}{character string specify the direction of fold changes ('\code{up}' (positive fold changes),
'\code{down}' (negative fold changes), \code{both}'(both positive and negative fold changes)), features changes
more than \code{abs.fold} in the defined direction are kept.}

\item{...}{additional arguments passing to \code{\link{rpkm}}, \code{\link{cpm}}}
}
\value{
A \code{TCA} object
}
\description{
This fuction constructs a time course table of which rows corrsponding to genomic features
and columns the timepoint. values can be mean normalized read counts or log2-fold changes compared
to the first timepoint. The time course table is used for clustering analysis.
}
\note{
If '\code{expression}' in \code{value} is chosen, for replicates , the normalized expression value is first
calculated for each replicate, then mean value is taken to represent the normalized expression value.
}
\examples{
data(tca_ATAC)
tca_ATAC <- DBanalysis(tca_ATAC)
tca_ATAC <- timecourseTable(tca_ATAC, value = 'expression', lib.norm = TRUE, norm.method = 'rpkm')

}
\author{
Mengjun Wu
}

