% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{clust-class}
\alias{clust}
\alias{clust-class}
\title{clust class}
\description{
\code{clust} is a S4 class for storing results of a clustering analysis for time course data.
}
\details{
:
The clust objects are returned from \code{\link{timeclust}}.

Slots in a clust object can be extracted by using \code{\link{@}} extraction operator.
}
\section{Slots}{

Oject of this class contains the following slots:
\describe{
 \item{\code{method}}{clustering method that has been used}
 \item{\code{dist}}{distance method that has been used}
 \item{\code{data}}{a matrix of original or standardized data that has been used for the analysis}
 \item{\code{centers}}{a matrix of class centers}
 \item{\code{cluster}}{an integer vector of length \eqn{n} (\eqn{n} is the number of data points
 each integer indicates the cluster a data point belongs to. For the fuzzy cmeans clustering method,
 a data point is assigned to the closest cluster to which the data point has highest membership value.}
 \item{\code{membership}}{a matrix with membership values of the data points to all the clusters}
}
}
\author{
Mengjun Wu
}
\seealso{
\code{\link{timeclust}}, \code{\link{@}}
}

