% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{count mapped reads overlap genomic intervals}
\usage{
countReads(object, dir, method = "summarizeoverlaps", ...)
}
\arguments{
\item{object}{a \code{TCA} object}

\item{dir}{character string giving the directory where BAM files are stored.}

\item{method}{character string giving the counting method. Options are '\code{summarizeOverlaps}'
and '\code{featureCounts}'. For Windows system, only '\code{summarizeOverlaps}' can be used, For Linux
system, both methods can be used.}

\item{...}{additional arguments passed to \code{\link{summarizeOverlaps}} and \code{featureCounts} in \code{Rsubread} package}
}
\value{
A TCA object with updated '\code{count}' slot.
}
\description{
This function counts mapped reads from multiple BAM files that overlap genomic intervals
in \code{genomicFeature} in a \code{TCA} object. The counting result is stored in '\code{count}'
slot of the \code{TCA} object.
}
\details{
This function provides two options '\code{summarizeOverlaps}' from GenomicAlignments package and '\code{featureCounts}'
from Rsubread package to count the aligned reads. As Rsubread package is only avaible for linux systems,
Windows users can only use '\code{summarizeOverlaps}'. The user could specify counting details by passing additional
arguments (...), otherwise the default settings of the two methods are used. For counting details, see
\code{\link{summarizeOverlaps}}, \code{featureCounts} in \code{Rsubread} package
}
\author{
Mengjun Wu
}
\seealso{
\code{\link{summarizeOverlaps}}, \code{featureCounts} in \code{Rsubread} package
}

