% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBresult.R
\name{DBresult}
\alias{DBresult}
\alias{DBresult.cluster}
\title{Perform likelihood ratio tests and extract the differential analysis results}
\usage{
DBresult(object, group1 = NULL, group2 = NULL, contrasts = NULL,
  p.adjust = "fdr")

DBresult.cluster(object, group1 = NULL, group2 = NULL, contrasts = NULL,
  p.adjust = "fdr", cluster, cmthreshold = NULL)
}
\arguments{
\item{object}{a \code{TCA} object, for \code{DBresult}, \code{DBanalysis} should already be called
on the object; for \code{DBresult.cluster}, both \code{DBanalysis} and \code{timeclust} should be
already called}

\item{group1}{character string giving the level to be compared, that is the denominator in the fold changes.}

\item{group2}{a character vetor giving other levels to compared with \code{group1}. that are numerator
in the fold changes.}

\item{contrasts}{a character vector, each charcter string in the vector gives a contrast of two groups with
the format 'group2vsgroup1', group1 is the denominator level in the fold changes and group2 is the numerator
level in the fold changes.}

\item{p.adjust}{character string specifying a correction method for p-values. Options are 'holm', 'hochberg',
'hommel', 'bonferroni', 'BH', 'BY', 'fdr', 'none'.}

\item{cluster}{an integer, the result tables of genomic features belong to
the \code{cluster} are extracted.}

\item{cmthreshold}{a numeric value, this argument is applicable only if '\code{cmeans}' clustering method
is selected when calling \code{\link{timeclust}} function. if not NULL, the result table of genomic features
that belong to the defined \code{cluster} and the membership values to this cluster exceed \code{cmthreshold}
are extracted.}
}
\value{
A named list of result tables for all provided contrasts, the names of the list are the contrasts. Each result table
is a data frame containing the following columns:

\code{logFC} log2-fold change of differential event between two tested.

\code{PValue} p-values.

\code{paj} adjusted p-values

\code{chr}  name of the chromosomes

\code{start} starting position of the feature in the chromosome

\code{end} ending postition of the feature in the chromosome

\code{id} genomic feature name
}
\description{
This function performs likelihood ratio tests for given coefficinets contrasts
after fitting read counts to GLM by \code{\link{DBanalysis}}. \code{DBresult} extracts
the result tables of differential analysis of given contrasts and the result tables contain
all genomic features. \code{DBresult.cluster} returns similar result tables while the result
tables only contains genomic features belong to a given cluster.
}
\details{
This function uses \code{\link{glmLRT}} from edgeR which perform likelihood ratio tests for testing
significance of changes. For more deatils, see \code{\link{glmLRT}}
}
\note{
If not NULL \code{group1}, \code{group2} and \code{contrasts}, result tables are extracted from comparisons
in \code{constrasts}.
}
\examples{
data(tca_ATAC)
tca_ATAC <- DBanalysis(tca_ATAC)
### extract differntial analysis of 24h, 72h to 0h
# set the contrasts using the 'group1' and 'group2' paramters
res1 <- DBresult(tca_ATAC, group1 = '0h', group2 = c('24h', '72h'))
# one can get the same result by setting the contrasts using hte 'contrasts' parameter
res2 <- DBresult(tca_ATAC, contrasts = c('24hvs0h', '72hvs0h'))

# extract differntial analysis of 24h, 72h to 0h of a given cluster
tca_ATAC <- timecourseTable(tca_ATAC, filter = TRUE)
tca_ATAC <- timeclust(tca_ATAC, algo = 'cm', k = 6)

}
\author{
Mengjun Wu, Lei Gu
}
\seealso{
\code{\link{glmLRT}}
}

